@@ -4,14 +4,14 @@ publish_to: 'none'
 version: 0.1.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   cupertino_icons: ^1.0.3
   flex_color_scheme: ^3.0.1
   flutter:
     sdk: flutter
-  flutter_riverpod: ^0.14.0+3
+  flutter_riverpod: ^1.0.0-dev.7
 
 dev_dependencies:
   flutter_test:
